requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Database
supabase==2.0.0
postgrest==0.13.0

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.13.0  # Compatible with Python 3.11
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2

# HTTP & API
httpx==0.25.2
requests==2.31.0

# Utils
pydantic==2.5.2
email-validator==2.1.0.post1
python-slugify==8.0.1

# Security
cryptography==41.0.7
PyJWT==2.8.0